LPD-81556 Moving DoAsUserThread from portal-impl to portal-test because it is only used for tests, and its code should not be in production code
