+++ portal-test/src/com/liferay/portal/test/security/permission/DoAsUserThread.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.security.permission;
+package com.liferay.portal.test.security.permission;
 
 import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.portal.kernel.log.Log;
